Move pytest fixtures to conftest.py
