+++ tests/test_candidate_graph/test_iou.py
@@ -1,53 +1,5 @@
-from motile_toolbox.candidate_graph.iou import compute_ious
 import pytest
-import numpy as np
-from skimage.draw import disk
-
-@pytest.fixture
-def segmentation_2d():
-    frame_shape = (100, 100)
-    total_shape = (2, *frame_shape)
-    segmentation = np.zeros(total_shape, dtype="int32")
-    # make frame with one cell in center with label 1
-    rr, cc = disk(center=(50, 50), radius=20, shape=(100, 100))
-    segmentation[0][rr, cc] = 1
-
-    # make frame with two cells
-    # first cell centered at (20, 80) with label 1
-    # second cell centered at (60, 45) with label 2
-    rr, cc = disk(center=(20, 80), radius=10, shape=frame_shape)
-    segmentation[1][rr, cc] = 1
-    rr, cc = disk(center=(60, 45), radius=15, shape=frame_shape)
-    segmentation[1][rr, cc] = 2
-
-    return segmentation
-
-def sphere(center, radius, shape):
-    assert len(center) == len(shape)
-    indices = np.moveaxis(np.indices(shape), 0, -1)  # last dim is the index
-    distance = np.linalg.norm(np.subtract(indices, np.asarray(center)), axis=-1)
-    mask = distance <= radius
-    return mask
-
-
-@pytest.fixture
-def segmentation_3d():
-    frame_shape = (100, 100, 100)
-    total_shape = (2, *frame_shape)
-    segmentation = np.zeros(total_shape, dtype="int32")
-    # make frame with one cell in center with label 1
-    mask = sphere(center=(50, 50, 50), radius=20, shape=frame_shape)
-    segmentation[0][mask] = 1
-
-    # make frame with two cells
-    # first cell centered at (20, 50, 80) with label 1
-    # second cell centered at (60, 50, 45) with label 2
-    mask = sphere(center=(20, 50, 80), radius=10, shape=frame_shape)
-    segmentation[1][mask] = 1
-    mask = sphere(center=(60, 50, 45), radius=15, shape=frame_shape)
-    segmentation[1][mask] = 2
-
-    return segmentation
+from motile_toolbox.candidate_graph.iou import compute_ious
 
 
 def test_compute_ious_2d(segmentation_2d):
